backend/requirements.txt
asgiref==3.7.2
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.1
deprecation==2.1.0
Django==5.0
django-cors-headers==4.3.1
django-keycloak-auth==0.9.9
django-uw-keycloak==2.0.2
djangorestframework==3.14.0
dry-rest-permissions==0.1.10
ecdsa==0.18.0
gunicorn==21.2.0
idna==3.6
jwcrypto==1.5.6
mysqlclient==2.2.0
packaging==23.2
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
python-jose==3.3.0
python-keycloak==3.9.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
